audio queue with interpolation

+++ Source/Synth/ChorusVoice.h
@@ -22,7 +22,7 @@ public:
 	virtual void ProcessBlock(AudioSampleBuffer & outputBuffer, int startSample, int numSamples) override;
 private:
 	std::unique_ptr<CAudioQueue> mBuffer;
-	std::vector<int> mVoiceOffsets;
+	std::vector<double> mVoiceOffsets;
 	IVoltageController& mPhaseLfo;
 	double mLengthSec;
 	double mWet = 0;
